move luminescence plate block to softmax pro structure
