+++ src/allotropy/parsers/moldev_softmax_pro/softmax_pro_structure.py
@@ -7,6 +7,13 @@ from typing import Any, Optional
 import uuid
 
 from allotropy.allotrope.allotrope import AllotropeConversionError
+from allotropy.allotrope.models.luminescence_benchling_2023_09_luminescence import (
+    DeviceControlAggregateDocument as DeviceControlAggregateDocumentLuminescence,
+    DeviceControlDocumentItem as DeviceControlDocumentItemLuminescence,
+    MeasurementAggregateDocument as MeasurementAggregateDocumentLuminescence,
+    MeasurementDocumentItem as MeasurementDocumentItemLuminescence,
+    Model as ModelLuminescence,
+)
 from allotropy.allotrope.models.shared.definitions.custom import (
     TQuantityValueDegreeCelsius,
     TQuantityValueNanometer,
@@ -21,10 +28,8 @@ from allotropy.allotrope.models.ultraviolet_absorbance_benchling_2023_09_ultravi
 )
 from allotropy.parsers.lines_reader import LinesReader
 from allotropy.parsers.moldev_softmax_pro.fluorescence_plate_block import (
+    FluorescenceOrLuminescencePlateBlock,
     FluorescencePlateBlock,
-)
-from allotropy.parsers.moldev_softmax_pro.luminescence_plate_block import (
-    LuminescencePlateBlock,
 )
 from allotropy.parsers.moldev_softmax_pro.plate_block import (
     Block,
@@ -42,6 +47,67 @@ from allotropy.parsers.utils.values import (
 
 BLOCKS_LINE_REGEX = r"^##BLOCKS=\s*(\d+)$"
 END_LINE_REGEX = "~End"
+
+
+class LuminescencePlateBlock(FluorescenceOrLuminescencePlateBlock):
+    READ_MODE = "Luminescence"
+    CONCEPT = "luminescence"
+    UNIT = "RLU"
+    DATA_TYPE_IDX = 6
+    EXCITATION_WAVELENGTHS_IDX = 19
+
+    def generate_device_control_doc(self) -> DeviceControlDocumentItemLuminescence:
+        device_control_doc = DeviceControlDocumentItemLuminescence(
+            detector_gain_setting=self.pmt_gain
+        )
+
+        if self.is_single_wavelength:
+            device_control_doc.detector_wavelength_setting = TQuantityValueNanometer(
+                self.wavelengths[0]
+            )
+
+        return device_control_doc
+
+    def generate_measurement_doc(
+        self, well: str, well_data: WellData
+    ) -> MeasurementDocumentItemLuminescence:
+        measurement = MeasurementDocumentItemLuminescence(
+            DeviceControlAggregateDocumentLuminescence(
+                [self.generate_device_control_doc()]
+            ),
+            self.generate_sample_document(well),
+        )
+
+        if well_data.temperature is not None:
+            measurement.compartment_temperature = TQuantityValueDegreeCelsius(
+                float(well_data.temperature)
+            )
+
+        if not well_data.is_empty:
+            measurement.data_cube = self.generate_data_cube(well_data)
+
+        if well_data.processed_data:
+            measurement.processed_data_aggregate_document = (
+                self.generate_processed_data_aggreate_document(well_data)
+            )
+
+        return measurement
+
+    def to_allotrope(self) -> Any:
+        wells = sorted(self.well_data.keys(), key=natural_sort_key)
+
+        allotrope_file = ModelLuminescence(
+            measurement_aggregate_document=MeasurementAggregateDocumentLuminescence(
+                measurement_identifier=str(uuid.uuid4()),
+                plate_well_count=TQuantityValueNumber(self.num_wells),
+                measurement_document=[
+                    self.generate_measurement_doc(well, self.well_data[well])
+                    for well in wells
+                ],
+            )
+        )
+
+        return allotrope_file
 
 
 class AbsorbancePlateBlock(PlateBlock):
